# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Momentum Terminal API",
    "description": "US equities 12-1 momentum research and INTERNAL paper simulation. No brokerage connectivity.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/rebalance/plans/{plan_id}": {
      "get": {
        "tags": [
          "rebalance"
        ],
        "summary": "Plan Detail",
        "operationId": "plan_detail_api_rebalance_plans__plan_id__get",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rebalance/plans/{plan_id}/apply": {
      "post": {
        "tags": [
          "rebalance"
        ],
        "summary": "Apply",
        "description": "Apply the plan to the INTERNAL VIRTUAL PORTFOLIO ONLY. Nothing is sent to any broker.",
        "operationId": "apply_api_rebalance_plans__plan_id__apply_post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rebalance/plans/{plan_id}/skip": {
      "post": {
        "tags": [
          "rebalance"
        ],
        "summary": "Skip",
        "operationId": "skip_api_rebalance_plans__plan_id__skip_post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkipRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FillModel": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "session": {
            "type": "string",
            "title": "Session"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "side": {
            "type": "string",
            "enum": [
              "buy",
              "sell"
            ],
            "title": "Side"
          },
          "shares": {
            "type": "integer",
            "title": "Shares"
          },
          "ref_price": {
            "type": "number",
            "title": "Ref Price"
          },
          "fill_price": {
            "type": "number",
            "title": "Fill Price"
          },
          "gross_value": {
            "type": "number",
            "title": "Gross Value"
          },
          "slippage_cost": {
            "type": "number",
            "title": "Slippage Cost"
          },
          "commission": {
            "type": "number",
            "title": "Commission"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session",
          "symbol",
          "side",
          "shares",
          "ref_price",
          "fill_price",
          "gross_value",
          "slippage_cost",
          "commission"
        ],
        "title": "FillModel"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlanDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "signal_session": {
            "type": "string",
            "title": "Signal Session"
          },
          "fill_session": {
            "type": "string",
            "title": "Fill Session"
          },
          "status": {
            "type": "string",
            "enum": [
              "proposed",
              "applied",
              "executed",
              "skipped",
              "blocked"
            ],
            "title": "Status"
          },
          "block_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Reason"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "executed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Executed At"
          },
          "selected_count": {
            "type": "integer",
            "title": "Selected Count"
          },
          "est_turnover": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Est Turnover"
          },
          "realized_turnover": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Realized Turnover"
          },
          "data_label": {
            "type": "string",
            "title": "Data Label"
          },
          "data_version": {
            "type": "string",
            "title": "Data Version"
          },
          "config": {
            "$ref": "#/components/schemas/StrategyConfig"
          },
          "estimate": {
            "additionalProperties": true,
            "type": "object",
            "title": "Estimate"
          },
          "execution": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/RuleCheck"
            },
            "type": "array",
            "title": "Checks"
          },
          "orders": {
            "items": {
              "$ref": "#/components/schemas/PlanOrderModel"
            },
            "type": "array",
            "title": "Orders"
          },
          "targets": {
            "items": {
              "$ref": "#/components/schemas/PlanTarget"
            },
            "type": "array",
            "title": "Targets"
          },
          "fills": {
            "items": {
              "$ref": "#/components/schemas/FillModel"
            },
            "type": "array",
            "title": "Fills"
          },
          "can_apply": {
            "type": "boolean",
            "title": "Can Apply"
          },
          "apply_disabled_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apply Disabled Reason"
          },
          "portfolio_as_of": {
            "type": "string",
            "title": "Portfolio As Of"
          }
        },
        "type": "object",
        "required": [
          "id",
          "signal_session",
          "fill_session",
          "status",
          "block_reason",
          "created_at",
          "decided_at",
          "executed_at",
          "selected_count",
          "est_turnover",
          "realized_turnover",
          "data_label",
          "data_version",
          "config",
          "estimate",
          "execution",
          "checks",
          "orders",
          "targets",
          "fills",
          "can_apply",
          "apply_disabled_reason",
          "portfolio_as_of"
        ],
        "title": "PlanDetail"
      },
      "PlanOrderModel": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "side": {
            "type": "string",
            "enum": [
              "buy",
              "sell"
            ],
            "title": "Side"
          },
          "current_shares": {
            "type": "integer",
            "title": "Current Shares"
          },
          "target_shares": {
            "type": "integer",
            "title": "Target Shares"
          },
          "est_shares": {
            "type": "integer",
            "title": "Est Shares"
          },
          "est_price": {
            "type": "number",
            "title": "Est Price"
          },
          "est_value": {
            "type": "number",
            "title": "Est Value"
          },
          "est_cost": {
            "type": "number",
            "title": "Est Cost"
          },
          "current_weight": {
            "type": "number",
            "title": "Current Weight"
          },
          "target_weight": {
            "type": "number",
            "title": "Target Weight"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "order_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Status"
          },
          "filled_shares": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filled Shares"
          },
          "order_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Reason"
          }
        },
        "type": "object",
        "required": [
          "id",
          "symbol",
          "side",
          "current_shares",
          "target_shares",
          "est_shares",
          "est_price",
          "est_value",
          "est_cost",
          "current_weight",
          "target_weight",
          "note"
        ],
        "title": "PlanOrderModel"
      },
...
      "PlanTarget": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "momentum": {
            "type": "number",
            "title": "Momentum"
          },
          "close_raw": {
            "type": "number",
            "title": "Close Raw"
          },
          "adv20": {
            "type": "number",
            "title": "Adv20"
          },
          "target_weight": {
            "type": "number",
            "title": "Target Weight"
          },
          "current_weight": {
            "type": "number",
            "title": "Current Weight"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "rank",
          "momentum",
          "close_raw",
          "adv20",
          "target_weight",
          "current_weight"
        ],
        "title": "PlanTarget"
      },
...
      "RuleCheck": {
        "properties": {
          "rule": {
            "type": "string",
            "title": "Rule"
          },
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "rule",
          "ok",
          "detail"
        ],
        "title": "RuleCheck"
      },
...
      "SkipRequest": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "title": "SkipRequest"
      },
...
      "StrategyConfig": {
        "properties": {
          "initial_capital": {
            "type": "number",
            "maximum": 10000000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Initial Capital",
            "description": "Starting virtual cash, USD",
            "default": 100000.0
          },
          "top_n": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Top N",
            "description": "Number of top-ranked eligible stocks to hold",
            "default": 50
          },
          "lookback_sessions": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 2.0,
            "title": "Lookback Sessions",
            "description": "Older momentum anchor: sessions before signal",
            "default": 252
          },
          "skip_sessions": {
            "type": "integer",
            "maximum": 252.0,
            "minimum": 0.0,
            "title": "Skip Sessions",
            "description": "Recent anchor: sessions before signal (skip month)",
            "default": 21
          },
          "min_history_sessions": {
            "type": "integer",
            "maximum": 2000.0,
            "minimum": 0.0,
            "title": "Min History Sessions",
            "description": "Valid bars required before the signal session",
            "default": 252
          },
          "min_price": {
            "type": "number",
            "minimum": 0.0,
            "title": "Min Price",
            "description": "Raw (unadjusted) close at the signal session must exceed this, USD",
            "default": 5.0
          },
          "adv_window": {
            "type": "integer",
            "maximum": 252.0,
            "minimum": 1.0,
            "title": "Adv Window",
            "description": "Sessions in the average dollar-volume window",
            "default": 20
          },
          "min_adv_usd": {
            "type": "number",
            "minimum": 0.0,
            "title": "Min Adv Usd",
            "description": "Minimum trailing average daily dollar volume, USD",
            "default": 5000000.0
          },
          "slippage_bps": {
            "type": "number",
            "maximum": 500.0,
            "minimum": 0.0,
            "title": "Slippage Bps",
            "description": "Assumed adverse slippage vs. the open, basis points",
            "default": 10.0
          },
          "commission_per_order": {
            "type": "number",
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Commission Per Order",
            "description": "Assumed flat commission per fill, USD",
            "default": 0.0
          },
          "commission_bps": {
            "type": "number",
            "maximum": 500.0,
            "minimum": 0.0,
            "title": "Commission Bps",
            "description": "Assumed commission as bps of traded value",
            "default": 0.0
          },
          "min_session_coverage": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Session Coverage",
            "description": "Share of the point-in-time universe that must have a bar at the signal session; otherwise the rebalance is blocked",
            "default": 0.9
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date",
            "description": "Backtest start (ISO date); ignored by the paper ledger"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date",
            "description": "Backtest end (ISO date); ignored by the paper ledger"
          },
          "benchmark_symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Benchmark Symbol",
            "description": "Benchmark; default = provider's SPY (or demo stand-in)"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "StrategyConfig",
        "description": "All tunable inputs of the 12-1 momentum strategy.\n\nDefaults are the documented v1 specification. Every backtest run and every\npaper-portfolio configuration version stores the full JSON of this model."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}